Rewrite pages in TS
